Correct typo: 50,000 points -> 10,000 points
I was going through the mammoth data for the UMAP / tSNE projections (https://github.com/PAIR-code/understanding-umap/tree/master/raw_data), and all the files I checked contain 10,000 points, and looks identical to the data in the article. I don't know if the encoded data contains more points, but my guess is this was just a typo.